added bug ownership, only bug's owner can remove/update it

+++ miss-bug-frontend/src/pages/BugIndex.jsx
@@ -21,11 +21,18 @@ export function BugIndex() {
         setBugs(bugs)
     }
 
-    async function onRemoveBug(bugId) {
+    async function onRemoveBug(bugId, bugCreatorId) {
+        const loggedInUserId = userService.getLoggedinUser()?._id
+
+        console.log(loggedInUserId, bugCreatorId)
+        if (loggedInUserId !== bugCreatorId && bugCreatorId !== ""){
+            showErrorMsg('Only this bug\'s creator can remove it')
+            return
+        }
         try {
             await bugService.remove(bugId)
             console.log('Deleted Succesfully!')
-            setBugs(prevBugs => prevBugs.filter((bug) => bug._id !== bugId))
+            setBugs(prevBugs => prevBugs.filter(bug => bug._id !== bugId))
             showSuccessMsg('Bug removed')
         } catch (err) {
             console.log('Error from onRemoveBug ->', err)
@@ -57,10 +64,16 @@ export function BugIndex() {
     }
 
     async function onEditBug(bug) {
+        const loggedInUserId = userService.getLoggedinUser()?._id
+
+        if (loggedInUserId !== bug.creator._id && bug.creator._id !== ""){
+            showErrorMsg('Only this bug\'s creator can update it')
+            return
+        }
+
         const severity = +prompt('New severity?')
         const bugToSave = { ...bug, severity }
         try {
-
             const savedBug = await bugService.save(bugToSave)
             console.log('Updated Bug:', savedBug)
             setBugs(prevBugs => prevBugs.map((currBug) =>
